Vision Nectec Emotion/fastAPI/requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
boto==2.49.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.6.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.65.3
flatbuffers==1.12
fsspec==0.9.0
future==0.18.2
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.36.1
h11==0.12.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.4.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==3.0.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
openpifpaf==0.12.10
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.1.2
prometheus-client==0.10.1
prompt-toolkit==3.0.17
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pysparkling==0.6.1
python-dateutil==2.8.1
python-json-logger==2.0.1
python-multipart==0.0.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==52.0.0
six==1.15.0
sklearn==0.0
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toolz==0.11.1
torch==1.8.1
torchvision==0.9.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.36.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
